Allow fixing credentials after offline login, #3888
Also improve progress display in password prompt dialog

+++ src/misc/credentials/CredentialsProvider.ts
@@ -133,8 +133,9 @@ export interface ICredentialsProvider {
 	/**
 	 * Deletes stored credentials with specified userId.
 	 * No-op if credentials are not there.
+	 * @param opts.deleteOfflineDb whether to delete offline database. Will delete by default.
 	 */
-	deleteByUserId(userId: Id): Promise<void>
+	deleteByUserId(userId: Id, opts?: {offlineDb: "delete" | "keep"}): Promise<void>
 
 	/**
 	 * Sets the credentials encryption mode, i.e. how the intermediate key used for encrypting credentials is protected.
@@ -226,13 +227,15 @@ export class CredentialsProvider implements ICredentialsProvider {
 		})
 	}
 
-	async deleteByUserId(userId: Id): Promise<void> {
+	async deleteByUserId(userId: Id, opts: {offlineDb: "delete" | "keep"} = {offlineDb: "delete"}): Promise<void> {
 		const event: CredentialsDeletedEvent = {
 			name: CREDENTIALS_DELETED_EVENT,
 			userId,
 		}
 		this.interWindowEventBus?.send(event)
-		await this.offlineDbFacade?.deleteDatabaseForUser(userId)
+		if (opts.offlineDb === "delete") {
+			await this.offlineDbFacade?.deleteDatabaseForUser(userId)
+		}
 		this.storage.deleteByUserId(userId)
 	}
